Linear Regression model with bugs
Started creating a linear regression model from a tutorial and modifying it to a local CSV for a crypto currency. The hope is to make a 24 hour predictor using hourly data.

Once that works we can hook in an API or use data we collect or we can modify it to base one coin's price on other coins. Ex. the price of Doge Coin based on the price of BTC ETH and other bigger coins. Doing the second one would be just more data collection stuff, but the second one will involve a NN and probably some DL things.

 I'm not sure which I want to do first.

Currently there is a bug with the data. To fix it, change the 'date' from an object to a string. Then, you'll need to move it into the feature columns.

You'll also need to have added tensorflow to imports at this point.